Add failurePolicy variable for MutatingWebhookConfiguration
In our case we have a fallback scenario when there are no environment
variables that define kafka same AZ brokers, meaning we can still
connect using a single URL with the list of them.

We also host some other pods that does not depend on the webhook in the
same namespace that use this configuration. So it's fine for us to have
failurePolicy set to `Ignore`, in fact absolutely neccessary as we had
minor downtimes due to webhook misbehaving.

Currently when I set webhook replicas to 0 I can't manage the pods that
does not depend on it, but running in the same namespace as
namespaceSelector value, `Ignore` policy fixes it.